Route protection added along with controller to clean up route files and make them lean - better represents the MVC architecture.
